# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for HRD Attendance System",
        "title": "HRD Attendance API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of users with filter, sorting, pagination, and dynamic includes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get All Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Role (admin, manager, employee)",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc/desc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Relations (comma separated: tenant,attendances,attendances.user)",
                        "name": "includes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.UserResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new user with role hierarchy (SuperAdmin can create any, Admin can create HR/Employee)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create User",
                "parameters": [
                    {
                        "description": "User Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get authenticated user profile from token (httpOnly cookie) with optional includes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get current user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Relations (comma separated: tenant,attendances)",
                        "name": "includes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/users/me/activities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get recent activities for the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user's recent activities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.RecentActivityResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/profile-photo": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update logged-in user's profile photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update Profile Photo",
                "parameters": [
                    {
                        "description": "Profile Photo Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateProfilePhotoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get single user detail with optional includes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User By ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relations (comma separated: tenant,attendances)",
                        "name": "includes",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/modelDto.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/modelDto.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.UpdateProfilePhotoRequest": {
            "type": "object",
            "required": [
                "media_url"
            ],
            "properties": {
                "media_url": {
                    "type": "string"
                }
            }
        },
...
        "model.AttendanceResponse": {
            "type": "object",
            "properties": {
                "clock_in_latitude": {
                    "type": "number"
                },
                "clock_in_longitude": {
                    "type": "number"
                },
                "clock_in_media_url": {
                    "type": "string"
                },
                "clock_in_time": {
                    "type": "string"
                },
                "clock_out_latitude": {
                    "type": "number"
                },
                "clock_out_longitude": {
                    "type": "number"
                },
                "clock_out_media_url": {
                    "type": "string"
                },
                "clock_out_time": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.AttendanceStatus"
                },
                "user": {
                    "$ref": "#/definitions/model.UserResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.AttendanceStatus": {
            "type": "string",
            "enum": [
                "working",
                "done",
                "late"
            ],
            "x-enum-varnames": [
                "StatusWorking",
                "StatusDone",
                "StatusLate"
            ]
        },
...
        "model.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "123456"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.RecentActivityResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.RoleResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Tenant": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tenant_settings": {
                    "$ref": "#/definitions/model.TenantSetting"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.TenantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "tenant_settings": {
                    "$ref": "#/definitions/model.TenantSetting"
                }
            }
        },
        "model.TenantSetting": {
            "type": "object",
            "properties": {
                "allow_multiple_check": {
                    "type": "boolean",
                    "example": false
                },
                "allow_remote": {
                    "type": "boolean",
                    "example": false
                },
                "clock_in_end_time": {
                    "type": "string",
                    "example": "09:00"
                },
                "clock_in_start_time": {
                    "type": "string",
                    "example": "07:00"
                },
                "clock_out_end_time": {
                    "type": "string",
                    "example": "23:00"
                },
                "clock_out_start_time": {
                    "type": "string",
                    "example": "16:00"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "late_after_minute": {
                    "type": "integer",
                    "example": 480
                },
                "max_radius_meter": {
                    "type": "number",
                    "example": 100
                },
                "office_latitude": {
                    "type": "number",
                    "example": -6.1339179
                },
                "office_longitude": {
                    "type": "number",
                    "example": 106.8329504
                },
                "require_location": {
                    "type": "boolean",
                    "example": true
                },
                "require_selfie": {
                    "type": "boolean",
                    "example": true
                },
                "tenant": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-04-07T13:21:24Z"
                }
            }
        },
...
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Jl. Sudirman No. 1"
                },
                "attendances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AttendanceResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "type": "string",
                    "example": "IT"
                },
                "email": {
                    "type": "string",
                    "example": "budi@company.com"
                },
                "employee_id": {
                    "type": "string",
                    "example": "FT-001"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/profile/budi.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                },
                "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                },
                "recent_activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RecentActivityResponse"
                    }
                },
                "role": {
                    "$ref": "#/definitions/model.RoleResponse"
                },
                "tenant": {
                    "$ref": "#/definitions/model.TenantResponse"
                },
                "tenant_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "modelDto.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}